pyproject.toml
[project]
name = "buddy-os"
version = "0.1.0"
description = "Model-agnostic AI Orchestrator with dynamic model discovery"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=0.0.14",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.31.0",
    "aiosqlite>=0.20.0",
    "duckdb>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "agents", "ui"]